requirement.txt

groq
requests
cohere
python-dotenv